+++ src/main/java/site/metacoding/firstapp/web/PostController.java
@@ -6,16 +6,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class PostController {
 	
-	@GetMapping("/post/detailForm")
-	public String 상세보기() {
-		System.out.println("post detail");
-		return "post/detailForm";
-	}
-
-	@GetMapping("/post/listForm")
+	@GetMapping("/category/listForm")
 	public String 목록보기() {
-		System.out.println("listForm1");
-		return "post/listForm";
+		return "category/listForm";
 	}
 
 }
